Raise control exception after tracking and publishing result
If the desired behaviour being compares is an expected exception, such as ValidationErrors in Django, then eagerly raising the error will skip the publish step and not allow you to track the experiment.

This does not change the behaviour of the exception being raised, only that the other administrative functionality such as timing and publishing results get to run before the exception is raised up.